PlayerSprite class is added.
Currently all sprites have rotation values, which is not needed for our player sprite. It is because we have seperate images for front, back, left, and right directions. I will remove player rotations later.

Also, I will work on the player moving and animations for the next step. I guess there will be many things to change on physics and rendering system.