Sound effect: Money, Punch
I also changed the money update part of Physics loop, since I found a bug that the money simply disappear and doesn't count up in some cases. The problem was in the fact that the bullet and money are both LevelSpriteObject class and money sometimes considered as a bullet inside the big Physics loop. Thus, player sometimes took damage when collecting money. I fixed it.